scholarmind/requirements.txt
endee
sentence-transformers
scikit-learn
google-generativeai
streamlit
PyPDF2
rich
numpy
